backend/requirements.txt
Django==4.2.28
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
boto3==1.37.11
django-storages==1.14.5
Pillow==11.1.0
gunicorn==23.0.0
whitenoise==6.9.0
